+++ CHANGELOG.md
@@ -4,10 +4,17 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
+## [0.5.0] - 2020-07-13
+
+- Added *UiAssetLoader* to load Ui assets to memory
+
+**Changed**:
+- Removed the *UiService* dependency from the *com.gamelovers.assetLoader*
+
 ## [0.4.0] - 2020-07-13
 
 **Changed**:
-- Removed the *UiService* dependency from the *Addressables*
+- Removed the *UiService* dependency from the *com.unity.addressables*
 - Modified the *UiService* to be testable and injectable into other systems
 
 ## [0.3.2] - 2020-04-18